move kosit validator to main project
