+++ src/pax.XRechnung.NET/pax.XRechnung.NET.csproj
@@ -9,12 +9,12 @@
     <PackageReadmeFile>README.md</PackageReadmeFile>
     <RepositoryUrl>https://github.com/ipax77/pax.XRechnung.NET</RepositoryUrl>
     <PackageTags>dotnet;XRechnung;import;export;validate</PackageTags>
-    <AssemblyVersion>0.2.0</AssemblyVersion>
+    <AssemblyVersion>0.2.1</AssemblyVersion>
     <Owners>Philipp Hetzner</Owners>
     <Authors>Philipp Hetzner</Authors>
-    <Version>0.2.0</Version>
-	  <FileVersion>0.2.0.0</FileVersion>
-    <PackageVersion>0.2.0</PackageVersion>
+    <Version>0.2.1</Version>
+	  <FileVersion>0.2.1.0</FileVersion>
+    <PackageVersion>0.2.1</PackageVersion>
     <PackageLicenseExpression>MIT</PackageLicenseExpression>    
     <EnforceCodeStyleInBuild>true</EnforceCodeStyleInBuild>
     <EnableNETAnalyzers>true</EnableNETAnalyzers>
@@ -95,7 +95,7 @@
     <EmbeddedResource Include="Resources/Schematrons/PEPPOL-EN16931-UBL.sch" />
   </ItemGroup>
 
-    <ItemGroup>
+  <ItemGroup>
     <None Include="..\..\README.md">
       <Pack>True</Pack>
       <PackagePath>.</PackagePath>
@@ -103,6 +103,7 @@
   </ItemGroup>
 
     <ItemGroup>
-      <ProjectReference Include="..\pax.XRechnung.NET.Validator\pax.XRechnung.NET.Validator.csproj" />
+      <PackageReference Include="HtmlAgilityPack" Version="1.12.1" />
+      <PackageReference Include="Microsoft.Extensions.Http" Version="9.0.4" />
     </ItemGroup>
 </Project>
